requirements.txt
plexapi>=4.15.0
python-Levenshtein>=0.21.0
fuzzywuzzy>=0.18.0